Moved all of the hooks into Hooks.cpp:
This change was added for easier logic split, so that we can clearly see what type of hooks has been applied in one global file.
