+++ YAGA/Readers/DialogReader.cpp
@@ -2,18 +2,8 @@
 // Union SOURCE file
 
 namespace GOTHIC_ENGINE {
-	// This hook allows reading of the selected dialog choice.
-	HOOK Ivk_HighlightSelected_Union PATCH(&zCViewDialogChoice::HighlightSelected, &zCViewDialogChoice::HighlightSelected_Union);
-	void zCViewDialogChoice::HighlightSelected_Union()
+	void ReadDialogChoice(zCViewDialogChoice* choice)
 	{
-		THISCALL(Ivk_HighlightSelected_Union)();
-		Read(string(zCViewDialogChoice::GetSelectedText()).AToW());
-	}
-	// This hook reads the default dialog choice when a dialog box is presented.
-	HOOK Ivk_StartSelection_Union PATCH(&zCViewDialogChoice::StartSelection, &zCViewDialogChoice::StartSelection_Union);
-	void zCViewDialogChoice::StartSelection_Union()
-	{
-		THISCALL(Ivk_StartSelection_Union)();
-		Read(string(zCViewDialogChoice::GetSelectedText()).AToW());
+		Read(string(choice->GetSelectedText()).AToW());
 	}
 }
